Great rework

+++ src/main/kotlin/domain/model/SensorDto.kt
@@ -3,4 +3,4 @@ package domain.model
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class SensorDto(val id: Int?, val vin: String?, val gasLiters: Int?, val publicKey: String?)
+data class SensorDto(val id: Int?, val vin: String?, val gasLiters: Float?, val publicKey: String?)
